add cached language context refresh and upgrade faqt
